+++ packages/node/src/utils/cosmos.ts
@@ -24,6 +24,7 @@ import {
   SubqlCosmosBlockFilter,
   SubqlCosmosTxFilter,
 } from '@subql/types-cosmos';
+import { isObjectLike } from 'lodash';
 import { isLong } from 'long';
 import { CosmosClient } from '../indexer/api.service';
 import { BlockContent } from '../indexer/types';
@@ -92,7 +93,8 @@ export function filterMessageData(
     filter.contractCall &&
     !(
       filter.contractCall === data.msg.decodedMsg.msg ||
-      filter.contractCall in data.msg.decodedMsg.msg
+      (isObjectLike(data.msg.decodedMsg.msg) &&
+        filter.contractCall in data.msg.decodedMsg.msg)
     )
   ) {
     return false;